Move avd locator tests
