+++ AndroidSdk.Tests/AvdManager_Tests.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
 using Xunit;
@@ -108,5 +109,42 @@ public class AvdManager_Tests : AvdManagerTestsBase, IClassFixture<AvdManager_Te
 		var avds = Sdk.AvdManager.ListAvds();
 		var avd = Assert.Single(avds);
 		Assert.Equal(TestAvdName, avd.Name, ignoreCase: true);
+	}
+	
+	[Fact(Skip = SkipOnCI)]
+	public void LocatedPath()
+	{
+		var l = new AvdLocator();
+		var p = l.Locate();
+
+		Assert.NotNull(p);
+		Assert.NotEmpty(p);
+	}
+
+	[Fact]
+	public void LocatedPathForEnvVar()
+	{
+		var oldHome = Environment.GetEnvironmentVariable("ANDROID_AVD_HOME");
+
+		var tempAvdPath = Path.Combine(Path.GetTempPath(), "AndroidSdk.Tests", nameof(AvdLocator_Tests), nameof(LocatedPathForEnvVar), "android-avd-home");
+		Directory.CreateDirectory(tempAvdPath);
+
+		try
+		{
+			Environment.SetEnvironmentVariable("ANDROID_AVD_HOME", tempAvdPath);
+
+			var l = new AvdLocator();
+			var p = l.Locate();
+
+			Assert.NotNull(p);
+			Assert.NotEmpty(p);
+
+			Assert.Equal(tempAvdPath, p[0].FullName);
+		}
+		finally
+		{
+			Environment.SetEnvironmentVariable("ANDROID_AVD_HOME", oldHome);
+			Directory.Delete(tempAvdPath, true);
+		}
 	}
 }
